fix: tidy up repo restructure
