+++ internal/resources/common/schema/graph_beta/device_management/configuration_policy_assignments.go
