+++ internal/resources/common/schema/graph_beta/device_management/platform_script_assignments.go
