+++ internal/resources/common/schema/graph_beta/device_management/role_definition_assignments.go
