+++ internal/resources/common/shared_models/graph_beta/device_management/role_definition_assignments.go
