+++ internal/resources/common/shared_models/graph_beta/device_management/settings_catalog_assignments.go
