+++ internal/resources/common/state/graph_beta/device_and_app_management/configuration_policy_settings.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 
 	"github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/normalize"
-	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_and_app_management"
+	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_management"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
