+++ internal/resources/common/validators/graph_beta/device_management/configuration_policy_settings.go
