+++ internal/resources/common/validators/graph_beta/device_management/device_configuration_assignment_test.go
@@ -3,7 +3,7 @@ package sharedValidators
 import (
 	"testing"
 
-	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_and_app_management"
+	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_management"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
