+++ internal/resources/common/validators/graph_beta/device_management/windows_update_assignment.go
