+++ internal/resources/device_management/graph_beta/macos_platform_script/validate_assignment.go
@@ -3,7 +3,7 @@ package graphBetaMacOSPlatformScript
 import (
 	"fmt"
 
-	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_and_app_management"
+	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_management"
 )
 
 // ValidateAssignmentConfiguration validates the assignment configuration
